Made the whole thing more secured.
Before this, the cache used to stay unless refreshed, so if someone logged out and someone else logged in without refresshing, the cache will show the previous users stuff. Which is a serious violation. So now i made the Querykey in a way that it must check who the user before showing anything.